+++ manage/feedback/index.php
@@ -23,11 +23,21 @@ include_once("../../connectDB.php");
             include_once("../nav.php");
             ?>
             <?php
-            if (isset($_GET["function"]) == "del") {
-                if (isset($_GET["id"])) {
-                    $id = $_GET["id"];
-                    mysqli_query($conn, "DELETE FROM feedback WHERE FeedID = '$id'");
-                }
+            if (isset($_POST['btnUpdateFeedback']) && $_POST['txtupdateFeed'] == 1) {
+                $id = $_POST['txtIDFeed'];
+                mysqli_query($conn, "UPDATE feedback SET state = 0 WHERE FeedID = '$id'") or die('Invalid query: ' . mysqli_error($conn));
+                echo '<meta http-equiv="refresh" content = "0; URL="/>';
+            }
+            if (isset($_POST['btnUpdateFeedback']) && $_POST['txtupdateFeed'] == 0) {
+                $id = $_POST['txtIDFeed'];
+                mysqli_query($conn, "UPDATE feedback SET state = 1 WHERE FeedID = '$id'") or die('Invalid query: ' . mysqli_error($conn));
+                echo '<meta http-equiv="refresh" content = "0; URL="/>';
+            }
+            // Delete feedback
+            if (isset($_GET["function"]) == "del" && isset($_GET["id"])) {
+                $id = $_GET["id"];
+                mysqli_query($conn, "DELETE FROM feedback WHERE FeedID = '$id'");
+                echo '<meta http-equiv="refresh" content = "0; URL="/>';
             }
             ?>
 
@@ -39,10 +49,10 @@ include_once("../../connectDB.php");
                             <tr class="text-center">
                                 <th><strong>No.</strong></th>
                                 <th><strong>Content</strong></th>
-                                <th><strong>State</strong></th>
                                 <th><strong>Send Date</strong></th>
                                 <th><strong>Username</strong></th>
-                                <th><strong>Product ID</strong></th>
+                                <th><strong>Product Name</strong></th>
+                                <th><strong>Image</strong></th>
                                 <th><strong>Show</strong></th>
                                 <th><strong>Delete</strong></th>
                             </tr>
@@ -51,7 +61,7 @@ include_once("../../connectDB.php");
                         <tbody>
                             <?php
                             $No = 1;
-                            $sq = "SELECT * FROM feedback ORDER BY sendDate DESC";
+                            $sq = "SELECT * FROM feedback f INNER JOIN product p ON f.ProID = p.ProID ORDER BY sendDate DESC";
                             $res = mysqli_query($conn, $sq);
                             if (!$res) {
                                 die('Invalid query: ' . mysqli_error($conn));
@@ -59,19 +69,37 @@ include_once("../../connectDB.php");
                             while ($row = mysqli_fetch_array($res, MYSQLI_ASSOC)) {
                             ?>
                                 <tr>
-                                    <td class="text-center"><?php echo $No ?></td>
-                                    <td><?php echo $row["Content"]; ?></td>
-                                    <td class="text-center"><?php echo $row["state"]; ?></td>
-                                    <td class="text-center"><?php echo $row["sendDate"]; ?></td>
-                                    <td><?php echo $row["Username"]; ?></td>
-                                    <td class="text-center"><?php echo $row["ProID"]; ?></td>
-                                    <td style='text-align:center'>
-                                        <a href="?page=update_cbshow&&id=<?php echo $row["FeedID"]; ?>">
-                                            <i class="bi bi-pen-fill" style="color: black;"></i>
-                                        </a>
+                                    <td class="text-center"><?= $No ?></td>
+                                    <td><?= $row["Content"]; ?></td>
+                                    <td class="text-center"><?= $row["sendDate"]; ?></td>
+                                    <td><?= $row["Username"]; ?></td>
+                                    <td class="text-center"><?= $row["ProName"]; ?></td>
+                                    <td class="text-center">
+                                        <img src='../../Product/<?= $row["Pro_image"] ?>' border='0' width="100" />
                                     </td>
                                     <td style='text-align:center'>
-                                        <a href="?page=feedback_management&&function=del&&id=<?php echo $row["FeedID"] ?>" onclick="return deleteConfirm()">
+                                        <form action="" method="POST">
+                                            <input type="hidden" id="txtIDFeed" name="txtIDFeed" value="<?= $row['FeedID']; ?>">
+                                            <input type="hidden" id="txtupdateFeed" name="txtupdateFeed" value="<?= $row['state']; ?>">
+                                            <?php
+                                            if ($row['state'] == 1) {
+                                            ?>
+                                                <button type="submit" class="btn btn-success" name="btnUpdateFeedback">
+                                                    <i class="bi bi-check-lg"></i>
+                                                </button>
+                                            <?php
+                                            } else {
+                                            ?>
+                                                <button type="submit" class="btn btn-danger" name="btnUpdateFeedback">
+                                                    <i class="bi bi-x-lg"></i>
+                                                </button>
+                                            <?php
+                                            }
+                                            ?>
+                                        </form>
+                                    </td>
+                                    <td style='text-align:center'>
+                                        <a href="?function=del&id=<?= $row["FeedID"] ?>" onclick="return deleteConfirm()">
                                             <i class="bi bi-trash-fill" style="color: red;"></i>
                                         </a>
                                     </td>
